requirements.txt
black
python-dotenv
isort
pyserial
psutil
pyyaml
tonberry
pywin32
